split off module generators
