+++ src/gt4py/backend/numpy_backend.py
@@ -28,6 +28,7 @@ from gt4py.utils.attrib import Set as SetOf
 from gt4py.utils.attrib import attribkwclass as attribclass
 from gt4py.utils.attrib import attribute
 
+from .module_generator import BaseModuleGenerator
 from .python_generator import PythonSourceGenerator
 
 
@@ -376,7 +377,7 @@ class NumPySourceGenerator(PythonSourceGenerator):
         return sources
 
 
-class NumPyModuleGenerator(gt_backend.BaseModuleGenerator):
+class NumPyModuleGenerator(BaseModuleGenerator):
     def __init__(self):
         super().__init__()
         self.source_generator = NumPySourceGenerator(
